+++ src/facts_experiment_builder/io/paths.py
@@ -45,3 +45,7 @@ class ExperimentPaths:
     @property
     def compose_path(self) -> Path:
         return self.experiment_dir / _COMPOSE_FILENAME
+
+def make_output_dir(experiment_paths: ExperimentPaths) -> None:
+    output_dir = experiment_paths.output_dir
+    output_dir.mkdir(parents=True)
